fix: refresh 401 handle fixed + videocall fixed + mocks changed

+++ src/shared/api/interceptors/response.ts
@@ -3,36 +3,47 @@ import { TokenService, UserService } from '../services';
 import { authHttp } from '../instance';
 import { isResponseRefresh } from '../utils';
 
+import { redirect } from 'react-router-dom';
+import { store } from '@/app/redux';
+import { login, logout } from '@/entities/auth';
+
 const refreshTokens = async () => {
-  const tokens = TokenService.get();
+  try {
+    const tokens = TokenService.get();
 
-  if (!tokens) return;
+    if (!tokens) return;
 
-  const { jwtToken, refreshToken } = tokens;
+    const { jwtToken, refreshToken } = tokens;
 
-  if (!jwtToken || !refreshToken) return;
+    if (!jwtToken || !refreshToken) return;
 
-  const response = await authHttp.post('/RefreshToken', {
-    jwtToken,
-    refreshToken,
-  });
+    const response = await authHttp.post('/RefreshToken', {
+      jwtToken,
+      refreshToken,
+    });
 
-  if (response.status === 401) {
-    TokenService.destroy();
-    UserService.destroy();
+    if (response.status === 200) {
+      TokenService.save(response.data.tokens);
+      UserService.save(response.data.user);
+      store.dispatch(login(response.data.user));
+    }
+
+    return response;
+  } catch (error) {
+    if (
+      axios.isAxiosError(error) &&
+      error.response &&
+      error.response.status === 401
+    ) {
+      TokenService.destroy();
+      UserService.destroy();
+      store.dispatch(logout());
+      redirect('/login');
+    }
   }
-
-  if (response.status === 200) {
-    TokenService.save(response.data.tokens);
-    UserService.save(response.data.user);
-  }
-
-  return response;
 };
 
-const retryRequestWithNewToken = async (
-  config: InternalAxiosRequestConfig
-) => {
+const retryRequestWithNewToken = async (config: InternalAxiosRequestConfig) => {
   const newToken = TokenService.get()?.jwtToken;
   if (newToken) {
     config.headers.Authorization = `Bearer ${newToken}`;
@@ -44,7 +55,7 @@ export const refreshTokensInterceptor = async (error: AxiosError) => {
   if (
     error.config &&
     error.response &&
-    error.response.status === 401 &&
+    error.response.status === 500 && // TODO 401
     !isResponseRefresh(error.config.url)
   ) {
     const response = await refreshTokens();
